+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/pipline/object_storage/ObjectLoaderPartLoaderStep.kt
@@ -16,7 +16,7 @@ import io.micronaut.context.annotation.Requires
 import jakarta.inject.Named
 import jakarta.inject.Singleton
 import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.atomic.AtomicLong
+import java.util.concurrent.atomic.AtomicInteger
 
 @Singleton
 @Requires(bean = ObjectLoader::class)
@@ -31,7 +31,8 @@ class ObjectLoaderPartLoaderStep(
     @Named("batchStateUpdateQueue") val batchQueue: QueueWriter<BatchUpdate>,
 ) : LoadPipelineStep {
     override val numWorkers: Int = loader.numUploadWorkers
-    private val streamCompletionMap = ConcurrentHashMap<DestinationStream.Descriptor, AtomicLong>()
+    private val streamCompletionMap =
+        ConcurrentHashMap<DestinationStream.Descriptor, AtomicInteger>()
 
     override fun taskForPartition(partition: Int): LoadPipelineStepTask<*, *, *, *, *> {
         return LoadPipelineStepTask(
